iot2aps-lambda/tests/requirements.txt
pytest
pytest-mock
boto3
python-dotenv
